introduce general type

+++ deepqmc/solver/solver_orbital.py
@@ -113,7 +113,7 @@ class SolverOrbital(SolverBase):
             cumulative_loss = 0
             for data in self.dataloader:
 
-                lpos = Variable(data).float()
+                lpos = Variable(data)
                 lpos.requires_grad = True
 
                 loss = self.loss(lpos)
